수정: RemoteStorageItem.port Long→Int, FTP 포트 하드코딩 제거, override 누락 수정
- RemoteStorageItem/RemoteStorageInfo port 타입 Long → Int
- RemoteStorageUseCase/StorageViewModel/AddRemoteStorageDialog 파라미터 동일하게 변경
- FtpFileSystem.connect()에서 item.port.toInt() → item.port (불필요한 변환 제거)
- ProgressTracker: progressMap/setProgressState/removeProgress에 override 추가
- DataStoreRepository: sortBy에 override 추가

+++ app/src/main/java/com/todokanai/composepracticenew/viewmodel/StorageViewModel.kt
@@ -49,7 +49,7 @@ class StorageViewModel @Inject constructor(
     }
 
     /** 원격 스토리지를 추가한다. */
-    fun addRemoteStorage(name: String, address: String, port: Long, id: String, password: String) {
+    fun addRemoteStorage(name: String, address: String, port: Int, id: String, password: String) {
         viewModelScope.launch {
             remoteStorageUseCase.add(name, address, port, id, password)
         }
